api/requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
black==21.12b0
celery==4.4.2
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-swagger==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flower==0.9.7
gunicorn==20.1.0
humanize==3.13.1
idna==2.10
intelli-sms-gateway==0.1.6
itypes==1.2.0
Jinja2==3.0.3
kombu==4.6.11
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
paynow==1.0.4
Pillow==8.3.1
platformdirs==2.4.0
prometheus-client==0.8.0
prompt-toolkit==3.0.28
# psycopg2==2.9.3
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyotp==2.6.0
pyparsing==3.0.7
pyPdf==1.13
PyPDF2==1.26.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.1.4
regex==2021.7.6
reportlab==3.5.34
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.2
tornado==6.1
typing-extensions==4.0.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.5
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.3.0
zipp==3.5.0
